Move long[] group varint methods to backward-codecs (#15113)
- Move deprecated long[] group varint methods from core to backward-codecs
- Remove GroupVIntUtil.readGroupVInts(DataInput, long[], int) from core
- Remove GroupVIntUtil.writeGroupVInts(DataOutput, byte[], long[], int) from core
- Remove DataOutput.writeGroupVInts(long[], int) from core
- Add GroupVIntUtil class in backward-codecs with long[] methods
- Add DataOutputUtil class in backward-codecs for long[] DataOutput operations
- Add comprehensive tests for the moved functionality
- Prevents accidental usage of legacy long[] methods in benchmarks
- Maintains backward compatibility for codecs that need long[] functionality

Fixes #15113
